services/gateway/requirements.txt
fastapi
uvicorn
pika
python-dotenv